+++ src/lib/plugins/dom-commands/buttons/index.ts
@@ -2,16 +2,10 @@ import { PluginRegistration } from '../../../base-iv';
 import { addButton, removeAllButtons, addButtonFactory, removeAllButtonsFactory, removeButtonFactory, AddRemoveAllButtons, AddAddButton } from './button-commands';
 
 export const buttonsPlugin: PluginRegistration = {
-  apiExtensions: [
-    {
-      apiName: 'addButton',
-      apiFn: addButton,
-    },
-    {
-      apiName: 'removeAllButtons',
-      apiFn: removeAllButtons,    
-    },
-  ],
+  apiExtension: {
+    addButton,
+    removeAllButtons
+  },
   targetFunctionFactories: [
     addButtonFactory,
     removeAllButtonsFactory,