+++ src/lib/plugins/general-commands/execution-modifiers/index.ts
@@ -11,31 +11,16 @@ const deprecate = (oldName: string, newName:string, fn: (...args: any[]) => any)
 };
 
 export const executionModifiersPlugin: PluginRegistration = {
-  apiExtensions: [{
-    apiName: 'return',
-    apiFn: deprecate('return', 'endAllNodes', stopExecution),
-  }, {
-    apiName: 'goSub',
-    apiFn: deprecate('goSub', 'runSync', runSync),
-  }, {
-    apiName: 'execute',
-    apiFn: deprecate('execute', 'runAsync', runAsync),
-  }, {
-    apiName: 'endAllNodes',
-    apiFn: stopExecution,
-  }, {
-    apiName: 'runSync',
-    apiFn: runSync,
-  }, {
-    apiName: 'runAsync',
-    apiFn: runAsync,
-  }, {
-    apiName: 'goToNode',
-    apiFn: goToNode,
-  }, {
-    apiName: 'wait',
-    apiFn: wait,
-  }],
+  apiExtension: {
+    return: deprecate('return', 'endAllNodes', stopExecution),
+    goSub: deprecate('goSub', 'runSync', runSync),
+    execute: deprecate('execute', 'runAsync', runAsync),
+    endAllNodes: stopExecution,
+    runSync,
+    runAsync,
+    goToNode,
+    wait,
+  },
   targetFunctionFactories: [
     stopExecutionFactory,
     pauseExecutionFactory,
@@ -48,5 +33,17 @@ export const executionModifiersPlugin: PluginRegistration = {
 
 declare module '../../../node' {
   interface NodeExtensions extends AddStopExecution, AddRunAsync, AddRunSync, AddGoToNode, AddWait {
+    /**
+     * Deprecated in favor of `endAllNodes()`
+     */
+    return: typeof stopExecution
+    /**
+     * Deprecated in favor of `runSync()`
+     */
+    goSub: typeof runSync
+    /**
+     * Deprecated in favor of `runAsync()`
+     */
+    execute: typeof runAsync
   }
 }